articles/examples/browser/requirements.txt
# AgentCore Browser Example
# https://docs.aws.amazon.com/bedrock-agentcore/

boto3>=1.34.0
python-dotenv>=1.0.0

# Optional: for browser automation
playwright>=1.40.0
